+++ source/react/components/Header.js
@@ -9,7 +9,7 @@ class Header extends React.Component {
             <Navbar brand='UCSD CSSA 短租平台' right>
                 <NavItem href='/#/login'>登陆</NavItem>
                 <NavItem href='/#/register'>注册</NavItem>
-                <NavItem href='/#/aboutus'>关于我们</NavItem>
+                <NavItem href='/#/about'>关于我们</NavItem>
                 <NavItem href='/#/publish'>发布</NavItem>
                 <NavItem href='/#/CheckInfo'>查看短租信息</NavItem>
                 <NavItem href='/#/ChangePassword'>修改密码</NavItem>
