+++ source/react/components/Main.js
@@ -14,11 +14,10 @@ import Login from './Login.js';
 import NotFound from './NotFound.js';
 import Publish from './Publish.js';
 import ChangePassword from './ChangePassword.js';
-import Register from './Register.js';
 import CheckInfo from './CheckInfo.js';
 import ChangeProfile from './ChangeProfile.js';
 import Register from './Register.js';
-import AboutUs from './AboutUs.js';
+import About from './About.js';
 
 //
 const Main = () => {
@@ -27,7 +26,7 @@ const Main = () => {
             <Route exact path='/' component={Home} />
             <Route exact path='/login' component={Login} />
             <Route exact path='/register' component={Register} />
-            <Route exact path='/aboutus' component={AboutUs} />
+            <Route exact path='/about' component={About} />
 
             <Route exact path='/publish' component={Publish} />
             <Route exact path='/checkInfo' component={CheckInfo} />
